# Conflicts:
#	fluss-common/src/main/java/com/alibaba/fluss/predicate/And.java
#	fluss-common/src/main/java/com/alibaba/fluss/predicate/CompareUtils.java
#	fluss-common/src/main/java/com/alibaba/fluss/predicate/CompoundPredicate.java
#	fluss-common/src/main/java/com/alibaba/fluss/predicate/Contains.java
#	fluss-common/src/main/java/com/alibaba/fluss/predicate/EndsWith.java
#	fluss-common/src/main/java/com/alibaba/fluss/predicate/Equal.java
#	fluss-common/src/main/java/com/alibaba/fluss/predicate/FieldRef.java
#	fluss-common/src/main/java/com/alibaba/fluss/predicate/FunctionVisitor.java
#	fluss-common/src/main/java/com/alibaba/fluss/predicate/GreaterOrEqual.java
#	fluss-common/src/main/java/com/alibaba/fluss/predicate/GreaterThan.java
#	fluss-common/src/main/java/com/alibaba/fluss/predicate/In.java
#	fluss-common/src/main/java/com/alibaba/fluss/predicate/IsNotNull.java
#	fluss-common/src/main/java/com/alibaba/fluss/predicate/IsNull.java
#	fluss-common/src/main/java/com/alibaba/fluss/predicate/LeafFunction.java
#	fluss-common/src/main/java/com/alibaba/fluss/predicate/LeafPredicate.java
#	fluss-common/src/main/java/com/alibaba/fluss/predicate/LeafUnaryFunction.java
#	fluss-common/src/main/java/com/alibaba/fluss/predicate/LessOrEqual.java
#	fluss-common/src/main/java/com/alibaba/fluss/predicate/LessThan.java
#	fluss-common/src/main/java/com/alibaba/fluss/predicate/NotEqual.java
#	fluss-common/src/main/java/com/alibaba/fluss/predicate/NotIn.java
#	fluss-common/src/main/java/com/alibaba/fluss/predicate/NullFalseLeafBinaryFunction.java
#	fluss-common/src/main/java/com/alibaba/fluss/predicate/Or.java
#	fluss-common/src/main/java/com/alibaba/fluss/predicate/PartitionPredicateVisitor.java
#	fluss-common/src/main/java/com/alibaba/fluss/predicate/Predicate.java
#	fluss-common/src/main/java/com/alibaba/fluss/predicate/PredicateBuilder.java
#	fluss-common/src/main/java/com/alibaba/fluss/predicate/PredicateVisitor.java
#	fluss-common/src/main/java/com/alibaba/fluss/predicate/StartsWith.java
#	fluss-common/src/test/java/com/alibaba/fluss/predicate/PredicateBuilderTest.java
#	fluss-common/src/test/java/com/alibaba/fluss/predicate/PredicateTest.java
#	fluss-flink/fluss-flink-common/src/main/java/com/alibaba/fluss/flink/source/FlinkSource.java
#	fluss-flink/fluss-flink-common/src/main/java/com/alibaba/fluss/flink/source/FlinkTableSource.java
#	fluss-flink/fluss-flink-common/src/main/java/com/alibaba/fluss/flink/source/PredicateConverter.java
#	fluss-flink/fluss-flink-common/src/main/java/com/alibaba/fluss/flink/source/enumerator/FlinkSourceEnumerator.java
#	fluss-flink/fluss-flink-common/src/main/java/com/alibaba/fluss/flink/utils/TypeUtils.java
#	fluss-flink/fluss-flink-common/src/test/java/com/alibaba/fluss/flink/source/enumerator/FlinkSourceEnumeratorTest.java
#	fluss-lakehouse/fluss-lakehouse-paimon/src/test/java/com/alibaba/fluss/lakehouse/paimon/flink/LakeTableEnumeratorTest.java

+++ fluss-common/src/test/java/com/alibaba/fluss/memory/TestingMemorySegmentPool.java
@@ -1,11 +1,12 @@
 /*
- * Copyright (c) 2024 Alibaba Group Holding Ltd.
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
+ *    http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -48,18 +49,18 @@ public class TestingMemorySegmentPool implements MemorySegmentPool {
     }
 
     @Override
-    public int totalSize() {
+    public long totalSize() {
         return pageSize;
     }
 
     @Override
     public void returnPage(MemorySegment segment) {
-        // do noting.
+        // do nothing.
     }
 
     @Override
     public void returnAll(List<MemorySegment> memory) {
-        // do noting.
+        // do nothing.
     }
 
     @Override
